+++ eva/src/components/dashBoardInside/dashChart.vue
@@ -23,7 +23,9 @@
       >
         <template v-for="(group, nGroup) in metricsByGroup">
           <div
-            v-for="(metric, nMetric) in group.filter(({hideLegend}) => !hideLegend)"
+            v-for="(metric, nMetric) in group
+              .filter(({hideLegend}) => !hideLegend)
+              .filter(({name}) => name !== xMetric)"
             :key="metric.name"
             @mouseenter="chart.highlightMetric(metric)"
             @click="openSettingsForMetric(metric, nGroup, nMetric)"
@@ -44,6 +46,12 @@
         class="svg-container"
         @dblclick="$emit('resetRange')"
       />
+      <div
+        v-show="dataRestFrom.length"
+        class="x-metric-text"
+      >
+        {{ xMetric }}
+      </div>
     </div>
     <DashChartSettings
       ref="chartSettings"
@@ -219,8 +227,13 @@ export default {
     },
 
     xMetric() {
-      const [xMetric] = this.firstDataRowMetricList;
-      return xMetric;
+      const { xAxis = {} } = this.options;
+      const { xMetric = null } = xAxis;
+      if (xMetric) {
+        return xMetric;
+      }
+      const [firstMetric] = this.firstDataRowMetricList;
+      return firstMetric;
     },
 
     box() {
@@ -228,7 +241,7 @@ export default {
       const { width, height } = sizeFrom;
       return {
         width: Math.round(width - 42),
-        height: Math.round(height - 55) - 30,
+        height: Math.round(height - 55) - 45,
       };
     },
 
@@ -483,6 +496,9 @@ export default {
 .svg-container
   position: relative
   min-height: 100px
+
+.x-metric-text
+  font-size: 11px
 
 .legend
   text-align: left